refactor: flatten module hierarchy and add consolidated minimal.ts
Flatten the src/functions directory by moving all module files up
one level into src/. This reduces nesting depth and simplifies
import paths across the codebase.

- move arithmetic, bitwise, combinatorics, construction, core,
  expression, fraction, geometry, logical, matrix, numeric,
  probability, relational, statistics, string, trigonometry,
  and utils modules
- update all import statements to reflect new direct paths
- add src/minimal.ts with a self-contained implementation of
  FixedPrecision, consolidating core functionality into a
  single file with no internal dependencies
