+++ src/arithmetic/power.ts
