+++ src/bitwise/operations.ts
