+++ src/combinatorics/operations.ts
