+++ src/core/context.ts
@@ -2,7 +2,7 @@ import type {
   FixedPrecisionConfig,
   FPContext,
   RoundingMode,
-} from "../../FixedPrecision";
+} from "../FixedPrecision";
 
 const ROUNDING_MODES = [0, 1, 2, 3, 4, 5, 6, 7, 8] as const;
