+++ src/matrix/operations.ts
