+++ src/numeric/constants.ts
