+++ src/probability/random.ts
